+++ GCR/composite.py
@@ -17,14 +17,19 @@ __all__ = ['CompositeSpecs', 'CompositeCatalog', 'MATCHING_FORMAT', 'MATCHING_OR
 MATCHING_FORMAT = None
 MATCHING_ORDER = tuple()
 
-
-def _match(index_this, index_main, sorter=None):
-    if sorter is None:
+def _match(index_this, index_main, sorter=None, secondary=None, sort_descending=True):
+    if (sorter is None) and (secondary is None):
         sorter = np.argsort(index_this)
+    elif sorter is None:
+        sorter = np.argsort(secondary)
+        if sort_descending:
+            sorter = sorter[::-1]
+        sorter2 = np.argsort(index_this[sorter], kind='stable')
+        sorter = sorter[sorter2]
     s = np.searchsorted(index_this, index_main, sorter=sorter)
     s[s >= len(sorter)] = -1
     matching_idx = sorter[s]
-    not_matched_mask = (index_this[matching_idx] != index_main)
+    not_matched_mask = (index_this[matching_idx] != index_main) 
     return matching_idx, not_matched_mask, sorter
 
 
@@ -32,7 +37,7 @@ def _slice_matched(input_data_dict, quantity_needed, matching_idx, not_matched_m
                    always_return_masked_array=False):
     for q in quantity_needed:
         data_this = input_data_dict[q][matching_idx]
-        if always_return_masked_array or not_matched_mask.any():
+        if always_return_masked_array or not_matched_mask.any() or type(not_matched_mask)==np.ma.core.MaskedArray:
             data_this = np.ma.array(data_this, mask=not_matched_mask)
         yield q, data_this
 
@@ -52,6 +57,11 @@ class CompositeSpecs(object):
     matching_by_column : str, optional (default: None)
         The column in this catalog to be used to match to the main catalog.
         If set, matching_row_order is assumed to be False.
+    secondary_sort_column: str, optional (default: None)
+        If set, the column to determine the preference between non-unique 
+        matches in this catalog. 
+    sort_descending: bool (default:True)
+        Whether a secondary sort should be performed in descending order
     matching_column_in_main : str, optional (default: None)
         The column in the main catalog to be used to match to this catalog.
     overwrite_quantities : bool (default: True)
@@ -70,6 +80,8 @@ class CompositeSpecs(object):
         matching_partition=True,
         matching_row_order=True,
         matching_by_column=None,
+        secondary_sort_column=None,
+        sort_descending=False,
         matching_column_in_main=None,
         overwrite_quantities=True,
         overwrite_attributes=True,
@@ -84,7 +96,7 @@ class CompositeSpecs(object):
         self.matching_row_order = bool(matching_row_order)
         self.matching_by_column = self.matching_column_in_main = None
         if matching_by_column:
-            self.set_matching_column(matching_by_column, matching_column_in_main, matching_partition)
+            self.set_matching_column(matching_by_column, matching_column_in_main, matching_partition, secondary_sort_column, sort_descending)
 
         # For backward compatibility
         if "matching_method" in kwargs:
@@ -94,7 +106,7 @@ class CompositeSpecs(object):
             elif matching_method == MATCHING_ORDER or matching_method == "MATCHING_ORDER":
                 self.set_matching_order()
             else:
-                self.set_matching_column(matching_method)
+                self.set_matching_column(matching_method, secondary_sort_column=secondary_sort_column, sort_descending=sort_descending)
 
         self.overwrite_quantities = bool(overwrite_quantities)
         self.overwrite_attributes = bool(overwrite_attributes)
@@ -140,11 +152,13 @@ class CompositeSpecs(object):
         self.matching_by_column = None
         self.matching_column_in_main = None
 
-    def set_matching_column(self, column, column_in_main=None, same_partition=False):
+    def set_matching_column(self, column, column_in_main=None, same_partition=False, secondary_sort_column=None, sort_descending=True):
         self.matching_partition = bool(same_partition)
         self.matching_row_order = False
         self.matching_by_column = column
         self.matching_column_in_main = column_in_main if column_in_main else column
+        self.secondary_sort_column = secondary_sort_column
+        self.sort_descending = sort_descending
 
     @property
     def is_valid_matching(self):
@@ -227,7 +241,7 @@ class CompositeCatalog(BaseGenericCatalog):
                 except (TypeError, KeyError, IndexError):
                     matching_method = MATCHING_FORMAT
                 cat = CompositeSpecs(instance, identifier, matching_method=matching_method)
-
+             
             self._catalogs.append(cat)
 
         # check uniqueness of main catalogs
@@ -307,6 +321,9 @@ class CompositeCatalog(BaseGenericCatalog):
             if cat.matching_by_column:
                 native_quantities_needed_dict[cat_id].add(cat.matching_by_column)
                 native_quantities_needed_dict[self._main.identifier].add(cat.matching_column_in_main)
+                if cat.secondary_sort_column:
+                    native_quantities_needed_dict[cat_id].add(cat.secondary_sort_column)
+
 
             # set order_matching_dummy_col if needed
             elif order_matching_dummy_col is None and cat.matching_row_order and not cat.matching_partition:
@@ -338,12 +355,21 @@ class CompositeCatalog(BaseGenericCatalog):
                 continue
 
             # match column if needed:
+
             if cat.matching_by_column:
-                matching_idx, not_matched_mask, cat.sorter = _match(
-                    cat.cache[cat.matching_by_column],
-                    data[(self._main.identifier, cat.matching_column_in_main)],
-                    cat.sorter,
-                )
+
+                if cat.secondary_sort_column:
+                    matching_idx, not_matched_mask, cat.sorter = _match(
+                        cat.cache[cat.matching_by_column],
+                        data[(self._main.identifier, cat.matching_column_in_main)],
+                        cat.sorter, cat.cache[cat.secondary_sort_column], cat.sort_descending
+                    )
+                else:
+                    matching_idx, not_matched_mask, cat.sorter = _match(
+                        cat.cache[cat.matching_by_column],
+                        data[(self._main.identifier, cat.matching_column_in_main)],
+                        cat.sorter,
+                    )
 
                 for q, v in _slice_matched(
                     cat.cache,
